+++ src/Errata.hs
@@ -1,5 +1,3 @@
-{-# LANGUAGE OverloadedStrings #-}
-
 {- |
 Module      : Errata
 Copyright   : (c) 2020 comp
@@ -17,6 +15,8 @@ To get started, see the documentation for 'prettyErrors'. When using this module
 The overall workflow to use the printer is to convert your error type to 'Errata', which entails filling in messages
 and 'Block's. You can create 'Errata' and 'Block' from their constructors, or use the convenience functions for
 common usecases, like 'errataSimple' and 'blockSimple'.
+
+For premade styles for blocks and pointers, take a look at "Errata.Styles".
 
 For easier reading, we define:
 
@@ -41,15 +41,11 @@ module Errata
     , Pointer (..)
       -- * Styling options
     , Style (..)
-    , basicStyle
-    , fancyStyle
-    , fancyRedStyle
-    , fancyYellowStyle
+    , PointerStyle (..)
       -- * Pretty printer
     , prettyErrors
     ) where
 
-import qualified Data.Text as T
 import qualified Data.Text.Lazy as TL
 import qualified Data.Text.Lazy.Builder as TB
 import           Errata.Internal.Render
@@ -70,66 +66,71 @@ errataSimple header block body = Errata
 
 -- | A simple block that points to only one line and optionally has a label, header, or body message.
 blockSimple
-    :: Style                               -- ^ The style of the pointer.
+    :: Style                               -- ^ The style of the block.
+    -> PointerStyle                        -- ^ The style of the pointer.
     -> FilePath                            -- ^ The filepath.
     -> Maybe Header                        -- ^ The header message.
     -> (Line, Column, Column, Maybe Label) -- ^ The line number and column span, starting at 1, and a label.
     -> Maybe Body                          -- ^ The body message.
     -> Block
-blockSimple style fp hm (l, cs, ce, lbl) bm = Block
+blockSimple style pstyle fp hm (l, cs, ce, lbl) bm = Block
     { blockStyle = style
     , blockLocation = (fp, l, cs)
     , blockHeader = hm
-    , blockPointers = [Pointer l cs ce False lbl]
+    , blockPointers = [Pointer l cs ce False lbl pstyle]
     , blockBody = bm
     }
 
 -- | A variant of 'blockSimple' that only points at one column.
 blockSimple'
-    :: Style                       -- ^ The style of the pointer.
+    :: Style                               -- ^ The style of the block.
+    -> PointerStyle                        -- ^ The style of the pointer.
     -> FilePath                    -- ^ The filepath.
     -> Maybe Header                -- ^ The header message.
     -> (Line, Column, Maybe Label) -- ^ The line number and column, starting at 1, and a label.
     -> Maybe Body                  -- ^ The body message.
     -> Block
-blockSimple' style fp hm (l, c, lbl) bm =
-    blockSimple style fp hm (l, c, c + 1, lbl) bm
+blockSimple' style pstyle fp hm (l, c, lbl) bm =
+    blockSimple style pstyle fp hm (l, c, c + 1, lbl) bm
 
 -- | A block that points to two parts of the source that are visually connected together.
 blockConnected
-    :: Style                               -- ^ The style of the pointer.
+    :: Style                               -- ^ The style of the block.
+    -> PointerStyle                        -- ^ The style of the pointer.
     -> FilePath                            -- ^ The filepath.
     -> Maybe Header                        -- ^ The header message.
     -> (Line, Column, Column, Maybe Label) -- ^ The first line number and column span, starting at 1, and a label.
     -> (Line, Column, Column, Maybe Label) -- ^ The second line number and column span, starting at 1, and a label.
     -> Maybe Body                          -- ^ The body message.
     -> Block
-blockConnected style fp hm (l1, cs1, ce1, lbl1) (l2, cs2, ce2, lbl2) bm = Block
+blockConnected style pstyle fp hm (l1, cs1, ce1, lbl1) (l2, cs2, ce2, lbl2) bm = Block
     { blockStyle = style
     , blockLocation = (fp, l1, cs1)
     , blockHeader = hm
-    , blockPointers = [Pointer l1 cs1 ce1 True lbl1, Pointer l2 cs2 ce2 True lbl2]
+    , blockPointers = [Pointer l1 cs1 ce1 True lbl1 pstyle, Pointer l2 cs2 ce2 True lbl2 pstyle]
     , blockBody = bm
     }
 
 -- | A variant of 'blockConnected' where the pointers point at only one column.
 blockConnected'
-    :: Style                       -- ^ The style of the pointer.
+    :: Style                               -- ^ The style of the block.
+    -> PointerStyle                        -- ^ The style of the pointer.
     -> FilePath                    -- ^ The filepath.
     -> Maybe Header                -- ^ The header message.
     -> (Line, Column, Maybe Label) -- ^ The first line number and column, starting at 1, and a label.
     -> (Line, Column, Maybe Label) -- ^ The second line number and column, starting at 1, and a label.
     -> Maybe Body                  -- ^ The body message.
     -> Block
-blockConnected' style fp hm (l1, c1, lbl1) (l2, c2, lbl2) bm =
-    blockConnected style fp hm (l1, c1, c1 + 1, lbl1) (l2, c2, c2 + 1, lbl2) bm
+blockConnected' style pstyle fp hm (l1, c1, lbl1) (l2, c2, lbl2) bm =
+    blockConnected style pstyle fp hm (l1, c1, c1 + 1, lbl1) (l2, c2, c2 + 1, lbl2) bm
 
 {- | A block that points to two parts of the source that are visually connected together.
 
 If the two parts of the source happen to be on the same line, the pointers are merged into one.
 -}
 blockMerged
-    :: Style                               -- ^ The style of the pointer.
+    :: Style                               -- ^ The style of the block.
+    -> PointerStyle                        -- ^ The style of the pointer.
     -> FilePath                            -- ^ The filepath.
     -> Maybe Header                        -- ^ The header message.
     -> (Line, Column, Column, Maybe Label) -- ^ The first line number and column span, starting at 1, and a label.
@@ -137,19 +138,20 @@ blockMerged
     -> Maybe Label                         -- ^ The label for when the two pointers are merged into one.
     -> Maybe Body                          -- ^ The body message.
     -> Block
-blockMerged style fp hm (l1, cs1, ce1, lbl1) (l2, cs2, ce2, lbl2) lbl bm = Block
+blockMerged style pstyle fp hm (l1, cs1, ce1, lbl1) (l2, cs2, ce2, lbl2) lbl bm = Block
     { blockStyle = style
     , blockLocation = (fp, l1, cs1)
     , blockHeader = hm
     , blockPointers = if l1 == l2
-        then [Pointer l1 cs1 ce2 False lbl]
-        else [Pointer l1 cs1 ce1 True lbl1, Pointer l2 cs2 ce2 True lbl2]
+        then [Pointer l1 cs1 ce2 False lbl pstyle]
+        else [Pointer l1 cs1 ce1 True lbl1 pstyle, Pointer l2 cs2 ce2 True lbl2 pstyle]
     , blockBody = bm
     }
 
 -- | A variant of 'blockMerged' where the pointers point at only one column.
 blockMerged'
-    :: Style                       -- ^ The style of the pointer.
+    :: Style                               -- ^ The style of the block.
+    -> PointerStyle                        -- ^ The style of the pointer.
     -> FilePath                    -- ^ The filepath.
     -> Maybe Header                -- ^ The header message.
     -> (Line, Column, Maybe Label) -- ^ The first line number and column, starting at 1, and a label.
@@ -157,128 +159,8 @@ blockMerged'
     -> Maybe Label                 -- ^ The label for when the two pointers are merged into one.
     -> Maybe Body                  -- ^ The body message.
     -> Block
-blockMerged' style fp hm (l1, c1, lbl1) (l2, c2, lbl2) lbl bm =
-    blockMerged style fp hm (l1, c1, c1 + 1, lbl1) (l2, c2, c2 + 1, lbl2) lbl bm
-
-{- | A basic style using only ASCII characters.
-
-Errors should look like so:
-
-> error header message
-> --> file.ext:1:16
-> block header message
->   |
-> 1 |   line 1 foo bar do
->   |  ________________^^ start label
-> 2 | | line 2
->   | |      ^ unconnected label
-> 3 | | line 3
->   | |______^ middle label
-> 4 | | line 4
-> 5 | | line 5
-> . | |
-> 7 | | line 7
-> 8 | | line 8 baz end
->   | |______^_____^^^ end label
->   |        |
->   |        | inner label
-> block body message
-> error body message
--}
-basicStyle :: Style
-basicStyle = Style
-    { styleLocation = \(fp, l, c) -> T.concat ["--> ", T.pack fp, ":", T.pack $ show l, ":", T.pack $ show c]
-    , styleNumber = T.pack . show
-    , styleLine = const id
-    , styleEllipsis = "."
-    , styleLinePrefix = "|"
-    , styleUnderline = "^"
-    , styleVertical = "|"
-    , styleHorizontal = "_"
-    , styleDownRight = " "
-    , styleUpRight = "|"
-    , styleUpDownRight = "|"
-    , styleTabWidth = 4
-    }
-
-{- | A fancy style using Unicode characters.
-
-Errors should look like so:
-
-> error header message
-> → file.ext:1:16
-> block header message
->   │
-> 1 │   line 1 foo bar do
->   │ ┌────────────────^^ start label
-> 2 │ │ line 2
->   │ │      ^ unconnected label
-> 3 │ │ line 3
->   │ ├──────^ middle label
-> 4 │ │ line 4
-> 5 │ │ line 5
-> . │ │
-> 7 │ │ line 7
-> 8 │ │ line 8 baz end
->   │ └──────^─────^^^ end label
->   │        │
->   │        └ inner label
--}
-fancyStyle :: Style
-fancyStyle = Style
-    { styleLocation = \(fp, l, c) -> T.concat
-        [ "→ ", T.pack fp, ":", T.pack $ show l, ":", T.pack $ show c
-        ]
-    , styleNumber = T.pack . show
-    , styleLine = const id
-    , styleEllipsis = "."
-    , styleLinePrefix = "│"
-    , styleUnderline = "^"
-    , styleHorizontal = "─"
-    , styleVertical = "│"
-    , styleDownRight = "┌"
-    , styleUpDownRight = "├"
-    , styleUpRight = "└"
-    , styleTabWidth = 4
-    }
-
--- | A fancy style using Unicode characters and ANSI colors, similar to 'fancyStyle'. Most things are colored red.
-fancyRedStyle :: Style
-fancyRedStyle = Style
-    { styleLocation = \(fp, l, c) -> T.concat
-        [ "\x1b[34m→\x1b[0m ", T.pack fp, ":", T.pack $ show l, ":", T.pack $ show c
-        ]
-    , styleNumber = T.pack . show
-    , styleLine = highlight "\x1b[31m" "\x1b[0m"
-    , styleEllipsis = "."
-    , styleLinePrefix = "\x1b[34m│\x1b[0m"
-    , styleUnderline = "\x1b[31m^\x1b[0m"
-    , styleHorizontal = "\x1b[31m─\x1b[0m"
-    , styleVertical = "\x1b[31m│\x1b[0m"
-    , styleDownRight = "\x1b[31m┌\x1b[0m"
-    , styleUpDownRight = "\x1b[31m├\x1b[0m"
-    , styleUpRight = "\x1b[31m└\x1b[0m"
-    , styleTabWidth = 4
-    }
-
--- | A fancy style using Unicode characters and ANSI colors, similar to 'fancyStyle'. Most things are colored yellow.
-fancyYellowStyle :: Style
-fancyYellowStyle = Style
-    { styleLocation = \(fp, l, c) -> T.concat
-        [ "\x1b[34m→\x1b[0m ", T.pack fp, ":", T.pack $ show l, ":", T.pack $ show c
-        ]
-    , styleNumber = T.pack . show
-    , styleLine = highlight "\x1b[33m" "\x1b[0m"
-    , styleEllipsis = "."
-    , styleLinePrefix = "\x1b[34m│\x1b[0m"
-    , styleUnderline = "\x1b[33m^\x1b[0m"
-    , styleHorizontal = "\x1b[33m─\x1b[0m"
-    , styleVertical = "\x1b[33m│\x1b[0m"
-    , styleDownRight = "\x1b[33m┌\x1b[0m"
-    , styleUpRight = "\x1b[33m└\x1b[0m"
-    , styleUpDownRight = "\x1b[33m├\x1b[0m"
-    , styleTabWidth = 4
-    }
+blockMerged' pstyle style fp hm (l1, c1, lbl1) (l2, c2, lbl2) lbl bm =
+    blockMerged pstyle style fp hm (l1, c1, c1 + 1, lbl1) (l2, c2, c2 + 1, lbl2) lbl bm
 
 {- | Pretty prints errors. The original source is required. Returns 'Data.Text.Lazy.Text' (lazy). If the list is empty,
 an empty string is returned.
@@ -304,7 +186,7 @@ toErrata :: ParseError -> 'Errata'
 toErrata (ParseError fp l c unexpected expected) =
     'errataSimple'
         (Just \"an error occured!\")
-        ('blockSimple' 'basicStyle' fp
+        ('blockSimple' 'Errata.Styles.basicStyle' 'Errata.Styles.basicPointer' fp
             (Just \"error: invalid syntax\")
             (l, c, c + T.length unexpected, Just \"this one\")
             (Just $ \"unexpected \" \<> unexpected \<> \"\\nexpected \" \<> T.intercalate \", \" expected))