+++ test/Errata/TypesSpec.hs
@@ -4,6 +4,7 @@ module Errata.TypesSpec
     ( spec
     ) where
 
+import Errata.Styles
 import Errata.Types
 import Test.Hspec
 
@@ -11,17 +12,20 @@ spec :: Spec
 spec = do
     describe "pointerColumns" $ do
         it "gets the columns" $
-            pointerColumns (Pointer 1 2 3 False Nothing) `shouldBe` (2, 3)
+            pointerColumns (Pointer 1 2 3 False Nothing basicPointer) `shouldBe` (2, 3)
 
     describe "highlight" $ do
         it "brackets parts of text" $
-            highlight "(" ")" [(2, 4), (6, 7)] "12345678" `shouldBe` "1(23)45(6)78"
+            highlight [(bracketPointer, (2, 4)), (bracketPointer, (6, 7))] "12345678" `shouldBe` "1(23)45(6)78"
 
         it "brackets empty column spans" $
-            highlight "(" ")" [(1, 1)] "1234" `shouldBe` "()1234"
+            highlight [(bracketPointer, (1, 1))] "1234" `shouldBe` "()1234"
 
         it "does nothing with no columns" $
-            highlight "(" ")" [] "123" `shouldBe` "123"
+            highlight [] "123" `shouldBe` "123"
 
         it "puts text for out of bounds columns at the end" $
-            highlight "(" ")" [(3, 5), (6, 7)] "1234" `shouldBe` "12(34)()"
+            highlight [(bracketPointer, (3, 5)), (bracketPointer, (6, 7))] "1234" `shouldBe` "12(34)()"
+
+bracketPointer :: PointerStyle
+bracketPointer = basicPointer { styleHighlight = \x -> "(" <> x <> ")" }