+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## Unreleased (0.4.0.0)
 
+* Added styling individual pointers with `PointerStyle`. This changes how `styleLine` and `highlight` works and moves to `styleUnderline` to `PointerStyle`.
+
+* Moved existing and added new premade styles to `Errata.Styles`. `Errata` no longer exports premade styles.
+
 * Added support for characters with different widths (full-width, combining characters, others). The cabal flag `usewcwidth` (default false) can be enabled to use the native `wcwidth` function.
 
 * Added support for replacing tabs with spaces with the `styleTabWidth` option (defaults to 4).